% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_egm_plot.R
\name{mod_egm_plot}
\alias{mod_egm_plot}
\title{EGM Plot module}
\arguments{
\item{id}{Module id}

\item{egm}{Reactive returning an EGM object}
}
\value{
None; used for side effects in Shiny
}
\description{
A Plotly WebGL viewer for EGM signal data.
}
\examples{
# In app_server: mod_egm_plot_server("plot", egm = uploaded$egm)
# In app_ui: mod_egm_plot_ui("plot")
}
